+++ packages/preview-middleware-client/src/adp/utils.ts
@@ -218,7 +218,7 @@ export async function getReuseComponentChecker(ui5VersionInfo: Ui5VersionInfo): 
         return reuseComponentChecker;
     }
 
-    let reuseComponentApi: IsReuseComponentApi;
+    let reuseComponentApi: typeof IsReuseComponentApi;
     if (!isLowerThanMinimalUi5Version(ui5VersionInfo, { major: 1, minor: 134 })) {
         reuseComponentApi = (await import('sap/ui/rta/util/isReuseComponent')).default;
     }
@@ -232,7 +232,7 @@ export async function getReuseComponentChecker(ui5VersionInfo: Ui5VersionInfo): 
         const component = FlexUtils.getComponentForControl(ui5Control);
 
         if (reuseComponentApi) {
-            return reuseComponentApi.isReuseComponent(component);
+            return reuseComponentApi(component);
         }
 
         if (!component) {