+++ packages/preview-middleware-client/types/sap.ui.rta.d.ts
@@ -80,6 +80,19 @@ declare module 'sap/ui/rta/plugin/AddXMLAtExtensionPoint' {
     }
 }
 
+declare module 'sap/ui/rta/plugin/AddXMLPlugin' {
+    import type CommandFactory from 'sap/ui/rta/command/CommandFactory';
+
+    interface Arguments {
+        commandFactory: CommandFactory;
+        fragmentHandler: (overlay: UI5Element, extensionPointInfo: uknown) => Promise<void | object>;
+    }
+
+    export default class AddXMLPlugin {
+        constructor(_: Arguments) {}
+    }
+}
+
 declare module 'sap/ui/rta/plugin/ExtendControllerPlugin' {
     import type CommandFactory from 'sap/ui/rta/command/CommandFactory';
 
@@ -155,10 +168,11 @@ declare module 'sap/ui/rta/command/OutlineService' {
 
 declare module 'sap/ui/fl/FakeLrepConnector' {
     export default class FakeLrepConnector {
-        static fileChangeRequestNotifier?: <T extends object>(
+        static fileChangeRequestNotifier?: <T extends object, U extends object>(
             fileName: string,
             kind: 'delete' | 'create',
-            change?: T
+            change?: T,
+            additionalChangeInfo?: U,
         ) => void;
         static enableFakeConnector: () => void;
     }
@@ -291,7 +305,7 @@ declare module 'sap/ui/rta/RuntimeAuthoring' {
         getService: <T>(name: 'outline' | 'controllerExtension' | string) => Promise<T>;
         getSelection: () => ElementOverlay[];
         getDefaultPlugins: () => { [key: string]: uknown; contextMenu: ContextMenu };
-        getPlugins: () => { [key: string]: unknown; contextMenu: ContextMenu };
+        getPlugins: () => { [key: string]: uknown; contextMenu: ContextMenu };
         setPlugins: (defaultPlugins: object) => void;
         getRootControlInstance: () => {
             getManifest(): Manifest;
@@ -323,12 +337,7 @@ declare module 'sap/ui/rta/util/hasStableId' {
 declare module 'sap/ui/rta/util/isReuseComponent' {
     import type Component from 'sap/ui/core/Component';
 
-    interface IsReuseComponentApi {
-        isReuseComponent(component?: Component): boolean;
-    }
-
-    const IsReuseComponentApi: IsReuseComponentApi;
-    export default IsReuseComponentApi;
+    export default function isReuseComponentApi(component: Component): boolean;
 }
 
 declare module 'sap/ui/rta/api/startAdaptation' {